ported: Testing sshd ported from the host system to the guest using statifier.sh
Since `sshd` expects `sshd-session` and `sshd-auth` to be located in the
`/usr/lib/ssh` directory, you can alter the behavior in sshd_config.

The following steps must be done manually:

Run `sshd-keygen -A` to generate the host key files.

`sshd` expect `usr/share/empty.sshd/` to be owned by root but the directory
itself actually created by uid of user (from host system), so we need to
manually chown it from the guest, although it would be better if there's
a way to fix this on host system.

You may need to bring up the loopback interface with `ip link set lo up`
in order to test `sshd` using an `nc` (netcat) client.

---

In most mainstream distros, `/etc/passwd` contains the character `x` in
the password field. In this setup, I omit it because the system does not
use `/etc/shadow`. This allows `sshd` to work without PAM, and no password
verification (e.g., `unix_chkpwd`) is required to log in as root, as long
as `PermitRootLogin` and `PermitEmptyPasswords` are set.

The `devpts` virtual filesystem is required for PTY creation, which is
necessary for user-space programs like `sshd`. While you can configure
the SSH client not to request a PTY with the `-T` option, this makes an
interactive remote shell impossible—you can only send a single command
before the connection closes.

Since `mount -t dev` overrides anything inside `/dev`, the creation of
`/dev/pts` has been moved from `start_vm.sh` to the init script.

---

Currently the system don't use `/etc/gshadow`.

This is the first time I know that another user can access privileges
owned by a specific group in which the user is not part of it by simply
remembering the group password. Wait, there's a password for the group?
I thought the password was only available for the user.

Signed-off-by: realyukii <reyuki@gnuweeb.org>