qemu: Use virtual disk image for persistent storage
The previous QEMU setup consisted only of a boot image, a Linux kernel executable,
and an init.cpio file (for initramfs, also known as early user space [1]).

This worked well for my typical “fire-and-forget” experiments and kernel
playground sessions. However, it would be useful if the ash shell history could
persist across runs. To achieve this, I modified `start_vm.sh` to let the user
choose between a temporary play-around mode and a persistent setup.

Unlike in early user space, I now need to specify the path to the init script
by passing `init` as a kernel parameter [2].

This change also removes all SSH-related lines from the configuration to
emphasize that the purpose of this repository is to provide a minimal,
usable Linux system. It’s up to the user to extend it beyond the provided
basic setup.

[1]:
- https://en.wikipedia.org/wiki/Booting_process_of_Linux#cite_ref-18
[2]:
- https://raw.githubusercontent.com/torvalds/linux/master/Documentation/admin-guide/kernel-parameters.txt

Signed-off-by: realyukii <reyuki@gnuweeb.org>